create videohash/tilemaker.py and updated videohash/videohash.py
There were some issues with only using whash for videos of samelength and kind of similar brighness levels.

Exmaples in https://github.com/akamhy/videohash/issues/88

To fix this issue we are making a image that is horizontally concatenation
of all the frames used for the collage in whash

Then we divide the horizontally concatenated into 64 equal tiles and
find the dominant color of the tile.

We have a deafult list of how we expect the dominant color to be, if they are
same as our default list we set the bit to one else set it to zero.

Finally take the bitwise XOR of the whash bit list and dominant color bitlist.